Add basic text parsing test case
