+++ runelite-client/src/main/java/net/runelite/client/plugins/shootingstars/StarRegion.java
@@ -1,4 +1,4 @@
-package net.runelite.client.plugins.starlink;
+package net.runelite.client.plugins.shootingstars;
 
 import java.util.Arrays;
 import java.util.List;
