+++ runelite-client/src/main/java/net/runelite/client/plugins/shootingstars/event/StarDowngradeEvent.java
@@ -1,0 +1,5 @@
+package net.runelite.client.plugins.shootingstars.event;
+
+public class StarDowngradeEvent
+{
+}
