+++ runelite-client/src/main/java/net/runelite/client/plugins/shootingstars/event/StarEvent.java
@@ -1,0 +1,8 @@
+package net.runelite.client.plugins.shootingstars.event;
+
+import net.runelite.client.plugins.shootingstars.EventType;
+
+public interface StarEvent
+{
+	EventType type();
+}
