--- runelite-client/src/main/java/net/runelite/client/plugins/starlink/event/StarDepletionEvent.java
@@ -1,5 +1,0 @@
-package net.runelite.client.plugins.starlink.event;
-
-public class StarDepletionEvent
-{
-}
