--- runelite-client/src/main/java/net/runelite/client/plugins/starlink/event/StarEvent.java
@@ -1,8 +1,0 @@
-package net.runelite.client.plugins.starlink.event;
-
-import net.runelite.client.plugins.starlink.EventType;
-
-public interface StarEvent
-{
-	EventType type();
-}
